+++ packages/player/src/internal/constants.ts
@@ -1,0 +1,6 @@
+export const mimeTypes = {
+  BTS: 'application/vnd.tidal.bts',
+  DASH: 'application/dash+xml',
+  EMU: 'application/vnd.tidal.emu',
+  HLS: 'application/vnd.apple.mpegurl',
+} as const;