+++ packages/player/src/player/browserPlayer.ts
@@ -356,6 +356,7 @@ export default class BrowserPlayer extends BasePlayer {
         playbackContext,
       },
     );
+    this.#isReset = false;
   }
 
   pause(): void {