+++ packages/auth/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3] - 2024-08-19
+
+### Changed
+
+- Omit scope validity check for client credentials
+
+## [1.3.2] - 2024-08-19
+
+### Changed
+
+- Remove scope param from client credentials request
+
 ## [1.3.1] - 2024-06-03
 
 ### Changed